This code will take random values 50 times and distribute among three entities such that in the end all are almost equal.
There are 3 boys Ram, Bheem, and Lakshman. All of them are to be always almost equally treated, based on the amount of cumulative award they are given.
We need a program to give award 50 times, based on a random draw. The draw determines the award value.
This award is then allocated to one of them.
While allocating the award you need to ensure it's given to the person who deserves it most, based on equal treatment